separated staff commands
